+++ src/main/resources/templates/employee/registration.html
@@ -22,7 +22,7 @@
 <div layout:fragment="~{content}">
   <div class="content">
     <h2>사원 등록</h2>
-    <form th:action="@{/employee/registration}" onsubmit="submitInsertForm(event)" enctype="multipart/form-data">
+    <form th:action="@{/api/employee/registration}" onsubmit="submitInsertForm(event)" enctype="multipart/form-data">
       <div class="form-group">
         <img id="employee-picture" class="employee-picture"/><br>
         <p>사진은 10MB까지 업로드 가능합니다.</p>
